IJPS-56: To fix search total count when no match found and code clean up

+++ src/main/java/uk/gov/hmcts/reform/jps/controllers/SittingRecordController.java
@@ -53,7 +53,7 @@ public class SittingRecordController {
 
         String hmctsServiceCode = Utility.validateServiceCode(requestHmctsServiceCode);
 
-        final int totalRecordCount = sittingRecordService.getTotalRecordCount(
+        final long totalRecordCount = sittingRecordService.getTotalRecordCount(
             sittingRecordSearchRequest,
             hmctsServiceCode
         );
